@naveen538 2019-09-03
job log:

Before patch :-
15:54:29 INFO : Running Host Tests Suite distro_stress_stress_stress
15:54:29 INFO : Running: /usr/bin/avocado run avocado-misc-tests/perf/stress.py -m /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/stress.py.data/stress.yaml --force-job-id 48f95c87ce7810e9de26a3408f1ec4b7d532f3aa --job-results-dir /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results
JOB ID : 48f95c87ce7810e9de26a3408f1ec4b7d532f3aa
JOB LOG : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-08-28T15.54-48f95c8/job.log
(1/4) avocado-misc-tests/perf/stress.py:Stress.test;default-file_size_per_thread-memory_per_thread-threads-71ef: ERROR: Failed to fetch stress-1.0.4.tar.gz. (0.94 s)
(2/4) avocado-misc-tests/perf/stress.py:Stress.test;default-file_size_per_thread-memory_per_thread-minimal-6f80: ERROR: Failed to fetch stress-1.0.4.tar.gz. (0.19 s)
(3/4) avocado-misc-tests/perf/stress.py:Stress.test;quick-file_size_per_thread-memory_per_thread-threads-1e5a: ERROR: Failed to fetch stress-1.0.4.tar.gz. (0.24 s)
(4/4) avocado-misc-tests/perf/stress.py:Stress.test;quick-file_size_per_thread-memory_per_thread-minimal-5626: ERROR: Failed to fetch stress-1.0.4.tar.gz. (0.25 s)
RESULTS : PASS 0 | ERROR 4 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME : 2.18 s
JOB HTML : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-08-28T15.54-48f95c8/results.html

After patch:-
avocado run stress.py -m stress.py.data/stress.yaml
JOB ID : f83ebcd127602c1b3b7a265b6ea5283295c42256
JOB LOG : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-08-30T15.06-f83ebcd/job.log
(1/4) stress.py:Stress.test;default-file_size_per_thread-memory_per_thread-threads-71ef: PASS (67.93 s)
(2/4) stress.py:Stress.test;default-file_size_per_thread-memory_per_thread-minimal-6f80: PASS (68.18 s)
(3/4) stress.py:Stress.test;quick-file_size_per_thread-memory_per_thread-threads-1e5a: PASS (19.30 s)
(4/4) stress.py:Stress.test;quick-file_size_per_thread-memory_per_thread-minimal-5626: PASS (13.83 s)
RESULTS : PASS 4 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME : 171.70 s
JOB HTML : /home/workspace/runAvocadoFVTTest/avocado-fvt-wrapper/results/job-2019-08-30T15.06-f83ebcd/results.html